Test version commit from version 4.26.2-15973114+++UE4+Release-4.26
This is a test to see whether my pc has the wrong version of unreal installed if so when downloading this branch my unreal wont launch and will give the usual message about incorrect versions. If i get this working i can install the same version on my pc and then have my github fully work. But for now i will have this separated to a different branch to cause no harm to the main project on my pc